"3:40pm. High waffling paper. Forget it. Let me check out the C++ library.
3:50pm. https://orbi.uliege.be/bitstream/2268/208421/1/Castronovo17-Offline_Policy-Search_BRL.pdf

Here is the PhD thesis by him.

4:20pm. I was wrong. It seems that Osband's paper was mostly a theory paper, and in fact...

http://proceedings.mlr.press/v87/morere18a/morere18a.pdf
Bayesian RL for Goal-Only Rewards

Something like this which uses Bayesian inference to update the posteriors is in fact the closest one could realistically get to Bayesian RL. This though occurred to me earlier, but I did not want to write it down till now.

Let me check out the C++ library and I will gather my thoughts.

This is difficult. I am not getting utility out of it.

4:25pm. Javascript beats C++. In this case hands down. The BBRL library has a dizzying amount of boilerplate everywhere.

I won't even bother studying this thing. Sometimes a simple glance at random segments is all it takes. I have no idea why the author picked C++ as his language of choice.

4:30pm. Somehow while reading Osband's paper, I got it into my head that there must be some kind of very simple tabular RL that just happens to replaces tabular values with conjugate distributions. On the virtue of being Bayesian it should supposedly work a lot better than the standard thing. Despite the underlying algorithm never being described, I sort of thought it unimaginably that he would write dozens and dozens of pages of proofs about without no computational basis underneath.

But by now I should know better how 'real' math works. It wasn't a coincidence that the tabular PSRL algorithm was never described.

I must have been wrong. I kind of feel bad about misleading the guys on the RL sub.

4:35pm. Let me read chapter 5 of that thesis which is on NNs.

After that I think I'll start to close the day with one of Wildberger's lectures. Then tomorrow I will track down some source code for MC-CFR and study that. I'll try translating the examples to F#, just like I did for the Java examples.

5pm. Done. Let me watch a video and then I will call it a day.

Today sucked. I keep doing these uninspiring excursions that waste time, and my focus has been low throughout. I am just going over things as if I am looking for something without knowing what. And somehow before I knew it the day is over.

I am going to do better once I get back into programming. It does not matter whether it is Python, or whether I build my own library in F# just for doing RL on the CPU. I will make it work without a shadow of a doubt this time.

Mastering MC-CFR will be the last bit of indulgence I will allow myself. Once I get back into it, I will do it all properly.

I will learn from my mistakes back in 2018 and will do a proper job as a researcher this time. I will create the agent and use it to smash my miserable zero net worth situation after that.

There are so, so many uses for RL, if deep nets could be made to work consistently.

My fate cannot possibly be to lose in such a dumb place as this. I cannot possibly get stuck here, right at the start of the race.

5:30pm. https://youtu.be/XXRwlo_MHnI?t=1104

Ah I had enough. 18m of this is the most I feel like today. I'll follow the schedule more exactly once I get into programming. Right now I still weaning myself off math.

It is really disappointing, but even though my goal was to be able to formalize anything, it is not my fault that I cannot do it. Rather the fields of mathematics are themselves to blame. I thought it might end up like this.

It is not that it is impossible, it is just that it feels like it would take so many man years to do it. Doable for the Inspired, but not so much for vanilla humans without access to the self improvement loop. I'll just have to go back to my old hit and miss methods. And hope that my studies of homotopy type theory give me hints as to what the further steps should be.

5:35pm. This is the only way things can go. Right now, even very simple algorithms are too difficult to reason about formally. It is kind of like the rationals vs repeating decimals situation. What I need is for type theory to allow me to transform the algorithms into a different kind of form that would allow for their elucidation.

Once that vital step is done the cruft of the olds fields could be dropped and the hit and miss methods abandoned.

But it is too difficult to just do this out of the blue. That sort of ideal can only be worked toward one step at a time.

And me, I am just weak, too weak. Even if I cannot go all the way, the possibility of having deep nets working satisfactorily is too enticing to pass up. Right now they have so many issue they cannot be considered to work even half right.

Just getting the barest of their basics right is too enticing for me to pass up. No matter how strong I think of myself, I just cannot resist the urge to run towards the oasis even if it might be a mirage. If it as all possible I'd upgrade my life status from dead to undead even if it is not quite live."