"9:05am. 4/13. I am looking at the Bayesian linear regression formulas in the `Bayesian RL for Goal-Only Rewards` paper and thinking, what if I replaced the covariance updates for this with the LSTD formulas? Wouldn't that make it work even better?
The paper claims that it combined the strengths of LSTD with uncertainty estimation, but I am not that sure. Let me figure out how LSTD works again.

9:20am. Ah, now I remember how I studied the Zap paper and figured out how the algorithm was derived. I completely forgot that I did that work. I remember explaining that one someone on Reddit last year. I should dig up that explanation for myself.

Damn it, I forgot so much. But I am convinced. There are all sorts of connections between these algorithms. Note that the covariance matrix estimate for Bayesian linear regression is the same as for KFAC. And the way the means are calculated are close to identical between the two methods. So I probably could replace it with the LSTD update and get an uncertainty estimate for the values.

I need to watch some lectures on how the updates for Bayesian linear regression are derived to make sure of that. The Wikipedia pages are useless as usual.

I could always ask the authors of the Zap paper about this as well, but I'll leave that thing for later when I am fully into it.

9:40am. I really need to start this thing instead of listening to Symphogear songs. So next week it ends, huh? The first two episodes were the best in the series - and Symphogear has very good initial episodes, but the rest was roughly on par.

I kind of feel bad that Shem-Ha will lose in the next episode. Imagine if she won and Symphogear S6 was about her war with the rest of the universe. That feels more interesting that anything we could get if the next season comes out normally. Her winning would be so completely beyond expectations that you'd know we are in a better timeline. I don't really identify with the protagonists.

9:50am. Let me watch a few Bayesian linear regression lectures. That is what I need to do. Then comes MC-CFR.

Since I haven't done real programming for so long, I feel out of practice. I'll get into the swing of things. This time I will figure it all out.

10am. https://www.youtube.com/watch?v=rVviNyIR-fI&list=PLD0F06AA0D2E8FFBA&index=52

Let me go through mathematicalmonk's lectures. I'll start with linear regression and work my way from there.

11:20am. https://youtu.be/n5QZL1B8FaE?list=PLD0F06AA0D2E8FFBA&t=183

I was bored for a while, but finally he is writing something that is new to me.

12:15pm. https://youtu.be/nrd4AnDLR3U?list=PLD0F06AA0D2E8FFBA

Let me stop here. Getting familiar with Bayesian linear regression should hopefully be the last bit of lectures before I start programming. Well, obviously, I'll have to study the MC-CFR papers in depth, but that is just a necessary part of it.

I am close to the starting line here.

With discrepancy reduction I have the missing glue I need to make everything work. It was the last piece that was missing. Once I start, I just have to setup the experiments properly and they will lead me to the finish line.

Long term, the subfields of ML will have to be restructured so they align with the developments in PL theory, but that is not a priority for the time being, and it is beyond my ability at the moment. I can only hope that this hit-and-miss style of making AI will suffice to get to the Singularity, but my luck does not feel that good."