"1:30pm. Let me resume. I've wasted enough time. Let me watch the Bayesian linear regression videos.
1:50pm. https://youtu.be/qz2U8coNwV4?list=PLD0F06AA0D2E8FFBA

He did go a bit quick near the end there. But I am motivated to learn this stuff in particular. It really makes a lot of difference between learning things at random, and learning things because they will directly help attain some goal.

2:15pm. No wait, wait. It is `A^T A` and not `1/m A^T A`?

Yes, this has been nagging me at the back of my head. The same goes for LSTD. But back in 2018 I normalized all the covariance matrices. What about Zap?

...Zap uses exponential averaging, just like KFAC.

Actually, I never realized this crucial difference between Zap and LSTD until now. I really have been wondering though how the posterior could narrow with more data if it always normalizes by the number of data points. I was also wondering why the rewards are just getting added in some of those formulas and never normalized.

So the similarities are only on the surface level.

2:35pm. https://youtu.be/xyuSiKXttxw?list=PLD0F06AA0D2E8FFBA

4 more to go. I really am learning quite a bit from this. I had no idea that linear regression used the pseudo-inverse.

3:15pm. https://youtu.be/LCISTY9S6SQ?list=PLD0F06AA0D2E8FFBA&t=610

I am taking a short breather.

http://ecmlpkdd2017.ijs.si/papers/paperID497.pdf
Bayesian Inference for Least Squares Temporal Difference Regularization

It seems somebody got around to this. I guess I do not have to ask Devray about it.

3:25pm. Uf, let me finish the last video and then I will read this paper. Then I can finally move on to MC-CFR.

3:40pm. https://youtu.be/g5BjTQghNf8?list=PLD0F06AA0D2E8FFBA&t=437

I am really not paying attention at this point. Somehow those earlier lectures sapped all my energy. I just do not give a shit anymore. Let me finish the lecture and then I will read the Bayesian LSTD paper.

3:45pm. https://youtu.be/g5BjTQghNf8?list=PLD0F06AA0D2E8FFBA&t=588

Gah, this uncertainty formula does not match the one in the paper. Why is it the inverse of the precision matrix. Should it not be the precision matrix?

Should `1 / lambda = 1 / a + x^T LAMBDA^-1 x` not be `1 / lambda = 1 / a + x^T LAMBDA x`?

3:55pm. Well, I will figure it out as I go along.

Actually, let me check out the PRML book for that formula and then I will check out the paper I just liked to. I can't believe the day is already nearly over.

4pm. According to PRML on page 174/748, the formula by mathematical monk is incorrect. I am not sure where he is getting that extra inverse. The EMU-Q paper is right.

No doubt, this should be put into the proof assistant and deal with that way.

4:05pm. Well, forget that for the time being.

4:50pm. Done with lunch. Let me resume. Right now I've just realized another difference between Zap and LSTD.

It seems that LSTD, takes reward directly as a target while Zap, takes the TD difference.

I can't believe I never spotted this.

Right now, one thing that is really bothering me is how to estimate aleatoric uncertainty.

If you take a look at ``1 / lambda = 1 / a + x^T LAMBDA x``, here in order to estimate uncertainty, it only ever looks at the inputs. Meaning reward scaling cannot possibly play into it.

In order to model aleatoric and epistemic uncertainty together, I think I am going to combine my Zap idea with the Gaussian sampling idea. I'll have to test to make sure that the Gaussian sampling idea actually does in fact model aleatoric uncertainty, but I will go with it in case it does. Alternatively, I can always go for quantile Q learning and find a way to combine it with Zap. But Gaussian sampling seems the most natural when combining models.

5:15pm. I reasoned out that it in fact does model aleatoric uncertainty last year by observing how the gradients behave for the sampling operation, but unfortunately, I forgot what my reasoning was and do not feel like going through the journal. I need to make sure of it experimentally anyway.

5:20pm. Let me go back to the paper I was reading. I had to take a short break after lunch.

```
The likelihood function for this model is given by:
p(b|θ, β) = N (b|Aθ, β^−1 C˜)
```

I admit that why the means formula is like this confuses me. Nonetheless, the way to track uncertainty for a given weight even in the Bayesian LSTD regime is unambiguous. It is the same as for regular Bayesian regression.

Admittedly this is confusing. Are `θ` really the weights?

`θ = (Φ˜^T (Φ˜ − γ Φ˜'))^−1 Φ˜^T R˜ = A^−1 b`

Ah, I see. So `Aθ` is just `b` which is `Φ˜^T R˜`.

5:30pm. "By setting the gradient equal to zero, we get the batch LSTD solution."

This sentence seems to be key. The equations seems much like what mathematicalmonk wrote, so it is not a stretch to think that the equations would turn out to be in alignment.

5:40pm. Ok, so the question of estimating parameter uncertainty by summing input vector covariances seems to be a sound one. Nonetheless, such an approach would not work for iterative updates.

Also I've come to a conclusion whether the targets should be whitened directly by multiplying the weights by the whitening matrix.

I think that the singularities involving invertibility would ruin that sort of thing. I'll be much better off sticking with KFAC simply because poor fitting would wreck the carefully trained weight matrix. It would only be good if there were no issues with invertibility which is never the case in practice.

So it is decided. Gassian sampling everywhere + ZAP in the last layer for the critic. KFAC for the actors. KFAC for the hidden layers as well.

But I won't touch multi-layer nets until I've gone through all the preliminaries. Unlike last time, this time I should make sure to master the linear players before I move on to the main course.

5:50pm. Ah, now I remember why I only decided to use sampling in the actors last year. If I did it in the value network as well, that would lead to exploration rewards being biased towards high aleatoric uncertainty areas.

So I cannot use sampling for the value network. Doh.

But in the case of huge rewards that would push the actors towards convergence prematurely, the modeling of aleatoric rewards induced by sampling would counterbalance that.

Now I understand it! I understand the reasoning that I took last year!

I really thought of everything back then when it come to picking an exploration method, haven't I?

I was just a tad short! If I had known about the principle of reversibility, I would not have had to throw in the towel.

6pm. Ah, damn it. Fine. Fine!

I just need to put a bit more into it. If it is just a poker player, I should have plenty of choices. It is not like I am trying to make Skynet at this point.

Just one success in order to establish a foothold.

Today I just ended up studying again, but tomorrow I see no reason why I should not dive directly into the MC-CFR paper. I already skimmed it once previously, but tomorrow I will seriously focus on it. Once I figure it out, I doubt it will be too hard to implement.

Actually, after I do that, I need to figure out what CFR+ is and such. I might as well make a real tour of it. It should not take too long.

6:05pm. It is difficult to describe the inner loneliness and weariness brought about by not being able to attain my goals for so long. I cannot deny the will, but I cannot deny the failure. So, so much time spent. A little success will clean it all out.

An improper victory will be inevitably follow by a loss. But a proper victory will have force and momentum behind it.

I am getting closer to it."