"1:25pm. Now where was I? I am just about ready to resume.
Lanctot's thesis.

http://mlanctot.info/

It is on his site.

"In this thesis, we investigate the problem of decision-making in large two-player zero-sum games using Monte Carlo sampling and regret minimization methods. We demonstrate four major contributions. The first is Monte Carlo Counterfactual Regret Minimization (MCCFR): a generic family of sample-based algorithms that compute near-optimal equilibrium strategies. Secondly, we develop a theory for applying counterfactual regret minimization to a generic subset of imperfect recall games as well as a lossy abstraction mechanism for reducing the size of very large games. Thirdly, we describe Monte Carlo Minimax Search (MCMS): an adversarial search algorithm based on *-Minimax that uses sparse sampling. We then present variance reduction techniques that can be used in these settings, with a focused application to Monte Carlo Tree Search (MCTS). We thoroughly evaluate our algorithms in practice using several different domains and sampling strategies."

I had no idea that MCCFR was introduced by him in this thesis.

It is 149 pages. This will take a while to go through.

1:55pm. "In this thesis, we focus on computing equilibria and decision-making in large games,
with perfect and imperfect information, sing Monte Carlo sampling algorithms."

`information, sing` should be `information, using`.

2:45pm. 19/149. Ah, let me take a break. Going through this in depth will take both time and focus as can be expected. But I'll master the sampling CFR algorithms by the end of it all. Then I'll be able to make poker agents on par with Pluribus. I am skeptical that they will be good enough to make money, but I suppose that view is one that needs testing.

3:05pm. Let me resume.

This will take a while. When I said, I would go through Lanctot's thesis, I sort of imagined it would take 5m and then I would get back to programming, but that sort of vision was too unrealistic. I guess going through this will take me a few days.

Hopefully I should have understanding of what this is about after that.

3:15pm. 21/149. I am really focusing on this part and I am actually internalizing the notation. This will be useful when I read the papers.

3:30pm. Focus me, focus. The thesis. Not /pol/.

4pm. 27/149. I am a bit confused over what the expectation is being taken here, but I am guessing over `u`s.

4:25pm. "To produce new strategies, regret-matching normalizes the positive portions of accumulated regret for each action [43; 42]. . More precisely, the strategy that regret-matching
produces at I for a ∈ A(I) is..."

```
let normalize array =
    let temp, normalizing_sum =
        Array.mapFold (fun s x ->
            let strategy = max x 0.0
            strategy, strategy + s
            ) 0.0 array

    let inline mutate_temp f = for i=0 to temp.Length-1 do temp.[i] <- f temp.[i]
    if normalizing_sum > 0.0 then mutate_temp (fun x -> x / normalizing_sum)
    else let value = 1.0 / float array.Length in mutate_temp (fun _ -> value)
    temp
```

Ah, so this strange normalization method is actually called regret matching.

4:45pm. 30/149. I understand everything on this page. It is pretty much how I did it in code except with a bit more mutation.

4:55pm. No wait, actually I am wrong. Note the lines 25 and onwards. I was wondering why the player swapping was not happening. It turns out, it does not update the second player at all. Hmmm...

Did I diverge from the Java examples on my own or were the examples that way from the start?

5:20pm. 34/149. I forgot what importance sampling was. I remember watching the mathematicalmonk lectures on this and it is still not coming back to me.

I am really getting tired at this point.

5:25pm. I am thinking of how to do the pruning optimization. I know where to put the cut.

```
            let util =
                if action_probability = 0.0 && two.probability = 0.0 then 0.0
                else next (action, {one with probability=one.probability*action_probability})
```

Here. This is how it should be. But I do not feel like testing this to make sure the optimization is sound.

5:50pm. 37/149. Let me stop here. Going through the entire thesis will take a while as expected, but I am in fact absorbing the material here.

It is not boring. The connection between Minimax and CFR is not something I realized until I read it just now. Though having said that, I am not familiar with Minimax so it is no wonder I failed to see the connection.

6pm. I can do this. Today I got the sense that I finally absorbed the notation, so the rest of the thesis and the VR-MCCFR paper (plus that other one Lanctot plugged me) should only take that much effort. Once I understand the algorithms it won't be too difficult to implement them for the toy games that I already have.

And whether it be Bluff or Liar Dice, the framework is flexible enough that the implementation of those won't take too much effort either. As long as none of the algorithms require complicated traversals like FSI-CFR I am safe."