"8:45am. I am going to paste a few snippets from the mail here.
```

> Why in the strategy update is the strategy weighted by the player's probability of reaching the infoset?

It's because that's how you compute the average strategy properly in extensive form games. (The average strategy is the one that converges to a Nash equilibrium over time).

If you want to dig deeper, you can look for "Kuhn's Theorem and the equivalence between mixed strategies and behavioral strategies".

> 2) Conversely, in the utility calculation, why - as seen from the start, is the utility weighed by just the opponent's probability of reaching the infoset and not both of them?

Ah yes, that is the definition of counterfactual value and counterfactual regret: it's exactly the novelty behind the CFR algorithm. It's because of this special form of regret that can do regret minimization in the extensive-form games and still get them to converge to the Nash equilibrium (via the CFR Theorem).  So the answer is essentially "because that's the definition" :)

> No matter how much I wrack my brain, I can't figure out the reasoning behind this.

There is a much more subtle reason that would only be clear to you if you go through all the CFR theorem proofs. The proofs simply cannot work unless you define regret in this way.
```

He also says that alternating updates CFR has better performance both in theory and in practice than the simultaneous ones. This is what I've observed yesterday. But finally, 'figure out the proofs' card got played against me.

8:50am. I've sensed that I am not going to be able to avoid this forever. Up to now I've been skipping proofs and I did not have the foundation to internalize them properly, but I am going to have to slow down and do this thing. This is why I've doing those Coq and Agda proofs for the past few months.

8:55am. I guess I will be stuck on this for a while. Let me take a look at that thing by Kuhn.

10:10am. 9/14 of 'Regret Minimization in Games with Incomplete Information'. In Eq 8 and 9 why are there two different definitions of the full counterfactual regret?

Let me fire a quick mail.

I also do not understand why `u(σ|D(I)→σ' , I') = u(σ|D(I')→σ' , I')`. Also why `succ(I'|I,a)` does not depend on the opponent's strategy is not explained.

11:05am. 'Define D(I) to be the information sets of player i reachable from I (including I).'

Ah, wait. I've been imaging this as only involving the immediate successors, but it should be more like all the way to the end. Why did I imagine only reaching the successors?

11:30am. Ok, I think I get theorem 3.

> I also do not understand why `u(σ|D(I)→σ' , I') = u(σ|D(I')→σ' , I')`.

It is not so much that `u(σ|D(I)→σ' , I') = u(σ|D(I')→σ' , I')`, but assuming that you have `u(σ|D(I)→σ' , I')` then you could probably construct `u(σ|D(I)→σ' , I') = u(σ|D(I')→σ' , I')`. So a more precise statement would be that `u(σ|D(I)→σ' , I')` implies `u(σ|D(I)→σ' , I') = u(σ|D(I')→σ' , I')`.

> Also why does `succ(I'|I,a)` not depend on the opponent's strategy?

I worded this badly. I meant to ask why it does not depend on the strategy of the player under consideration - meaning not the opponent, but the current player. And to answer this question, it is because `a` and `o'` are being maximized, so their probabilities of being picked will always be 1.0. Would that be right?

I guess this gives me a hint why CFR optimization does not consider the probabilities of the player being optimized.

> In Eq 8 and 9 why are there two different definitions of the full counterfactual regret?

I am guessing this is how pen and paper math goes. This is the first real proof I've ever tried to study. For the past few months as a part of my training I only did Coq and Agda proofs. I wonder if anybody tried putting the stuff these papers into a proof assistant?

...Is what I want to write...but that can't be right. Let me just add the first two. I still need for him to clarify eq 9 for me more.

I still do not understand why those two definitions of the full counterfactual regret in eq 8 and eq 9 are supposed to be equal. I am also not sure that the direction of the less-than-equals sign in eq 10 makes sense.

12pm. Let me take a break here. I've really been fully immersed into this proof. I guess him telling me to just study the thing is enough to serve as a trigger. Wonderful.

I really hope he is up for answering my questions because I do not think I will manage to get through the proof on my own."