"1:10pm. Done with breakfast, but let me chill a while longer.
1:30pm. For a bit longer. After that I will resume in full force.

1:55pm. Done with the break.

https://www.youtube.com/watch?v=KzZPueNROz8
The Tea Time Talks: Kris De Asis, Finite-Horizon Temporal Difference Methods (June 5)

This caught my interest so let me go through it. After that I'll resume studying the proofs.

2:15pm. https://youtu.be/KzZPueNROz8?t=818

Actually you could divide the horizons by two and then the learning would be sublinear in 2^(-H).

This idea could be promising, but regardless, using Zap + discrepancy reduction would be more promising.

I wish I'd thought of this idea myself though.

https://youtu.be/KzZPueNROz8?t=1142

Yeah, this is what one would expect.

2:35pm. This is such a simple and elegant idea. I love it.

2:45pm. Well, forget it. My idea is even better. Let me go back to figuring out CFR first.

2:50pm. Okay, I can agree that Lemma 5 is right baring the holes and slight logical inconsistencies. I understand what the full regret is, so saying that the full regret is <= immediate regret + full regret of successors is probably true. Right now I do not like the proof at all though. Unless I can get an explanation I will just pretend it is not there. Let me move on to the next thing.

Now that I focused on the definitions more closely though, I can see why the action probabilities of player under consideration are not used. Because the regret is minimized against the idealized simulacrum which has pure strategies only.

For what I can tell using the probabilities of both does work as one might expect, but it has a lot more variance. This just goes to show that if you are going to pick a target, then it should be an ideal one.

3pm. That having said, that only explains why the player one probabilities are not considered. Why do the player two probabilities have to play a role?

3:10pm. Node sharing purposes. I've been thinking about these lines. For example, if there are chance nodes that are not uniform you'd want to reflect that. For a game tree where there is no sharing, regular Minimax would do, but once sharing enters the picture to attain integrity, you'd want to multiply by the probabilities.

3:15pm. Ok, I had suspected that much yesterday.

Still, what confused me is that by my estimate, I am pretty sure that Dudo should have no sharing apart from the uniform chance nodes. That means that the Minimax style of updating should have worked and yet, the results were clearly wrong. The variance absolutely exploded in a way I cannot explain.

3:40pm. Focus me, focus. It is already this late. Where does time leave? I have no idea.

10/14. My focus is getting lower. It was high during the first proof, but since I could not internalize it I've gotten dejected.

I need to figure out how they go from the regret matching proof, to an iterative proof over the whole tree.

4:15pm. No I can't see it. I admit I've been skimming the later sections rather than full out focusing like for the first proof, but from what I can tell, they just generalize the Blackwell theorem and then using the theorem that `full regret <= sums of immediate regrets`, they conclude the case. But this is the issue for me, I am really having trouble figuring out what a system of mutually interactive nodes is doing. By comparison, what the individual nodes do is trivial. One can use the softmax or one can do regret matching, and both would work. I admit that out of everything that could be done, regret matching makes the most sense.

Is there a way to put a bound on `full regret <= sums of immediate regrets`? That would be interesting. That `<=` tells me absolutely nothing about the differential.

4:25pm. The next thing on the menu would be `Kuhn's Theorem and the equivalence between mixed strategies and behavioral strategies`. Forget about regret matching for a bit.

I agree with the basic logic that the less you have to multiply from the lower nodes, the better. The less unnecessary dependencies the better. And I agree that the environmental probabilities should be taken into consideration. And I agree that adding a dependency on the self when optimizing the self would make the variance go higher.

Let's say I understand those aspects of CFR.

I still do not understand the way the average strategy is computed at all. I cannot think about that in terms of optimization.

Maybe the reason why that is so confusing is because I am trying to connect it to the regret matching parts, but the strategy averaging is in fact a completely separate mechanism?

4:30pm. Say you were doing sampling - without all the corrections in outcome sampling. Then the frequency of updates of the average strategy would be dependent on the number of time it has been used.

...But that would not just depend on player one, but also on player two's probabilities.

Hmmm...I have no idea. No idea.

4:35pm. But I guess I am slowly getting closer to getting some sort of understanding. When I first started, I did not at all see the distinction between strategy averaging and the rest of the algorithm. Or rather, I saw it, thought it was very strange, but still considered that there is a dependency between regret matching and it. But now I am pretty sure there isn't one. Strategy averaging seems to be more like a Bayesian thing than anything else. Regret matching is optimization, but strategy averaging is not.

I should just dive into that paper, but I do not feel like it. Let me do the chores and then I will call it a day.

Maybe I'll do more once I get back."