"9:25am. I am up.
I finally figured out why there is not a multiply with `one_probability` when doing regret matching..

```
add node.regret_sum (fun i -> two_probability * (util.[i] - util_weighted_sum))
```

The above could be rewritten like...

```
add node.regret_sum (fun i -> (1.0 / float (actions.Length)) * two_probability * (util.[i] - util_weighted_sum))
```

And it would be equivalent. It it not that the dynamics on the player under consideration are not considered. It is just assumed that the sampling distribution is uniform.

This explains a few things.

1) It explains why the algorithm still converges after I put in `one_probability *`. I just end up replacing the uniform exploration method with a poorer one.

2) It explains why removing `two_probability` causes the algorithm to break completely. It is because it ends up assuming that the opposing player is playing uniformly. That is a much worse assumption than having a poor exploration method.

What CFR is doing is emulating a sampling algorithm in a deterministic manner. Those multiplies by the previous probabilities emulate the dynamics of it.

9:35am. Well, I'll ignore that the probabilities from the top do not match the dynamics.

The way CFR works is definitely interesting. Now that I've figured out this much, I can finally start figuring out what strategy averaging is going.

The fact that it multiplies only by the current player means it is assuming that the opponent is playing uniformly. I do not get that at all. What the hell? I am going to have to figure that out.

9:40am. Since I was too busy playing XCOM yesterday, let me watch Symphogear and then I will start. I do not want to leave it for later."