"10:15am. Let me finally start.
Throughout my life, when I consumed fiction I always looked for hints how to make the character powers realistic. Whether it be Symphogear or something else, right now nothing really comes close to my desire of showing truth about power. Deep down, I am really hungry for realism about the transcendent.

Simulacrum was a step forward, but even if I were to continue it, I have no idea what powers would be realistic. I want to know at any cost. No matter how much media comes out, it will never come close. I need to do it myself.

Killing yourself for power is cool, and radical when it was introduced, but I am satisfied just saying that it is viable. I do not want Simulacrum to just be about proselyting that view. I want to be the one to bring it closer to reality.

10:20am. To do that I need to get closer to the essence of AI.

I haven't gotten any reply from Lanctot yet. To be honest, I am starting to wonder whether he can even answer the questions about the proof given that he could not point out that player one's sampling distribution is uniform in CFR and just hid behind the proof and the definition. Or maybe just does not feel like it. Well, it is not like he has any obligation to do so, but it is a bad look for the field if it turns out that the very first proof makes no sense.

The point of them is not for the authors to look smart, but to convince me of something. So far it is utterly failing at that.

It just reinforces Wildberger's view that mathematical rigor has steadily gone down since Euclid's time.

Now that I understand formal theorem proving, unlike before I have a taste of what a proof is and is not. I know what I proof is and I have the ability to pass judgment. The ideal proof is like reading a step by step manual about something. The RL proofs that I've seen fall far short of that ideal.

And ideally, one would gather up the proofs from the papers and work on putting them into a proof assistant. This is more like a PhD project, and I do not want to pursue it personally.

10:25am. Leaving proof issues aside, the CFR algorithm itself is definitely worth studying. It is really a challenge to figure out what it does given all the different dynamics criss-crossing about.

I need to disentangle them.

Today I will work on figuring out how strategy averaging works. Suppose you were playing RPS, then it makes sense to not consider the probabilities of the other player when updating the strategy.

Still, I have no idea how to generalize this understanding to more complex games.

10:45am. Focus me. Let me finally start reading the paper.

11:30am. ...The paper is useless. It is unreadable in its complexity.

11:45am. "There is a well-known connection between regret, average strategies, and Nash equilibria."

Yeah, what is that connection? The one time I need a reference it is not there.

https://pdfs.semanticscholar.org/0184/855c7baafdbcadcab967d4bfa7d4f8b86285.pdf

"For each individual iteration of our training, the regrets may be temporarily skewed in such a way that an important strategy in the mix has a negative regret sum and would never be chosen. Regret sums and thus individual iteration strategies are highly erratic. What converges to a minimal regret strategy is the average strategy across all iterations. This is computed in a manner similar to getStrategy above, but without the need to be concerned with negative values."

12:10pm. https://int8.io/counterfactual-regret-minimization-for-poker-ai/#Why_Nash_Equilibrium

Maybe I am getting too pessimistic and impatient. Maybe I should just sit down and read the paper on Kuhn's theorem, but let me take a look at the alternatives first.

12:05pm. https://pdfs.semanticscholar.org/0184/855c7baafdbcadcab967d4bfa7d4f8b86285.pdf

The Java Intro to CFR just links to the original CFR paper, but does not explain why the average strategy converges to an equilibrium.

12:10pm. https://aaai.org/ocs/index.php/WS/AAAIW15/paper/viewFile/10110/10196

Let me read this thing.

12:20pm. 4/6. This paper is blisteringly complex. I can't understand it as I do not understand the methods being compared to.

12:30pm. Let me stop here. I am not as irritated as before, but I am not making much progress. I'll read a few papers and then just mail Zinkevich to ask him about strategy averaging, and maybe the proofs in the appendix."