"2:05pm. Almost done. Let me chill for a bit longer and then I will start.
2:25pm. https://arxiv.org/abs/1407.5042
Solving Large Imperfect Information Games Using CFR+

This is the paper that introduced CFR+. I'll leave it for later.

2:30pm. http://www.ii.uni.wroc.pl/~lukstafi/pmwiki/uploads/AGT/Prediction_Learning_and_Games.pdf

Somebody posted this book in a Reddit thread. Let me take a look.

I am looking up resources for no-regret learning.

2:35pm. Rather than make that mail later, let me do it now.

Maybe Lanctot will give me a reply in which case I'll be using up time of both people, but I have no intention of of pestering them for every little thing. I'll get these basic questions out of the way and then do the rest on my own unless it is something vital.

3pm.

"Hello. I have a few questions about the CFR algorithm and the paper.

It took me a while to figure out why in the regret matching equation, only the probability of the opposing player is used. At first I thought that the current player's probability was just dropped, but then I realized that this dropping off is equivalent to multiply by an uniform probability in each branch.

So for updating regrets, what CFR is doing is assuming that the current player is using an uniform strategy for exploration - at least in the preceding nodes. This is what imagine an equivalent sampling algorithm would be doing. The depth first search is essentially emulating a sampling algorithm in a deterministic manner.

1) That having said, I do not at all see the reason behind the average strategy updates. It assumes the the current player is following the strategy based off the current regrets, but that the opponent is using an uniform strategy.

I guess not considering opponent probabilities does make sense on toy games like rock paper scissors, but I can't at all figure why that reasoning would be valid in extended form games.

2) Why does strategy averaging work? Is there a simple explanation of this. I got plugged the 'Kuhn's theorem and the equivalence between mixed strategies and behavioral strategies' paper, but I find it too complicated an answer for what is in my mind a simple question.

I wish the paper went into depth on this.

3) I am not having much success understanding the first proof in the appendix. Taken at its face, the lemma that full regret is <= that the sums of individual regrets makes sense.

But in eq 8 the full counterfactual regret is defined, and then it is redefined to be something else in eq 9.

Is this supposed to mean that eq 8 and 9 are supposed to be equal? If so, what are the steps to go from 8 to 9?

Similarly between 9 and 10, the max is split and the equality because an <=. How does that work logically?

- Marko Grdinić"

Let me go with this. I am looking at his homepage and it has not been updated for 8 years. I hope he is still alive.

3:05pm. Now where was I? Let me check out that book. After that I'll go through the CFR+ paper.

3:15pm. I really hope he does not might me posting these in the journal.

"Hi Marko,

I have not been through this in a while, but I'm quite sure that Eq. 9 should be an inequality (less than or equal to) so not equivalent to Eq 8. This is because all it's doing is putting a max a \in A(I); it's upper-bounding the full counterfactual regret by saying it's the regret at I plus the regret of the subtree with the largest regret under it. This is in general looser than the mixed strategy (Eq 8) which does not necessarily put all of its weight on the maximal one.

succ(I' | I,a) certainly includes the other player's policy, because it depends on how the other players act at the intermediate nodes. Note that the definition has \sigma in the superscript, i.e. it's succ^{\sigma}(I' | I, a).

Hope this helps,
Marc"

I am sorry for doubting you. I guess you were busy. Yeah, I am nasty today.

"Thank you for the clarification. So assuming that eq 9 is an inequality, what exactly is going on going from eq 9 to 10 on a logical basis? How can the max operation be split as if it were distributive?

I know that I am nitpicking about details - sorry about that, but the only kind of theorem proving that I understand is the kind that is done in theorem provers like Coq or Agda. I want to see if I could at least attain an understanding of the first proof.

The lemma that the full regret can be split into sums of individual regrets makes sense to me, but when I try to cache myself into the proof, on eq 10 I realize that this split of the max might result in the inequality flipping. I just can't ground it on a logical basis and I can't see what in the structure of the equation 9 makes such an operation viable."

3:25pm. Let me get back to reading the book.

3:40pm. Actually let me take a break. Somehow I am completely drained. Unlike yesterday it does not feel like I did much of anything today.

4:15pm.

```
> Not sure why you'd think this step would flip the inequality...?

`max_{a} (f_1(a) + f_2(a))` restricts the action passed to both functions to be equal while optimizing them like in `max_a(f_1(a)) + max_a(f_2(a))` means the actions can be independent.  

Let me demonstrate with a bit of F# code.

---
type Actions = A | B | C

let f = function
    | A -> 1
    | B -> 0
    | C -> -100

let g = function
    | A -> -100
    | B -> 0
    | C -> 1

let max_comb = [|f A + g A; f B + g B; f C + g C|]
let max_indep = [|f A + g B; f A + g C; f B + g A; f B + g C; f C + g A; f C + g B|]

---

val max_comb : int [] = [|-99; 0; -99|]
val max_indep : int [] = [|1; 2; -100; 1; -200; -100|]

---

When I run the script as you can see from `max_comb` and `max_indep`, optimizing the components independently might result in better results than optimizing them together. When optimizing the max together the best solution is `f B + g B = 1` while for independent optimization the best solution is `f A + g B = 2`.

This is why I can't follow the step from 9 -> 10.
```

I wrote this out and then realized that this exactly proves what he said. Ah damn it. I've been making a mental error all this time. Somehow I've been under the impression that since the subject is regret minimization that the max should be minimizing the quantity. It just slipped past me somehow.

Gah! So it makes perfect sense.

4:55pm. Had to take some time to make a reply. Also finished studying the 6 page CFR+ paper. It is really quite a simple algorithm.

Revisiting CFR+ and Alternating Updates
https://arxiv.org/abs/1810.11542

5:15pm. Efficient CFR for Imperfect Information Games with Instant Updates
https://realworld-sdm.github.io/paper/27.pdf

I think I am going to call it a day here. I do not feel like doing it anymore. My focus is near its lowest point and has been for a while now.

Today it feels like all the real work was done during the night while I was in bed. I've only been gathering study materials while in front of the screen. I got that book, but I could not manifest even an iota of focus.

Tomorrow what I will do is take some time off to do the monthly review.

Then I will focus on going through that 400 page book on no regret learning. Hopefully I will figure out why the strategy averaging works after that. Hopefully that will improve my understanding and confidence enough that I will be able to return to implementing the sampling algorithm from Lanctot's thesis after that."